Put CRO inside nmf.py
